birthday calculator project
in this project the student will be able to:
1- Dealing with dates.
 2- Dealing with numbers.
 3- Dealing with advanced data types, Sequence, and their functions.
 4- Dealing with functions.

Requirements:
Build a program that takes in an unlimited number of names of people of their age. The program calculates the age, receives the date, and returns the current age, as well as the name of the day that he was born. The program also prints the age of the oldest and youngest person, and the total number of people passed through. If only one person is passed, the oldest and youngest person is not counted, and the phrase "There is no oldest or youngest person is printed". The date value is also validated, i.e., that the year, month, and day values must be valid (e.g., the 14 month value is invalid), and that they contain no symbols or negative values.